fix: a lot of naming + corrected the link btw studio and house
