+++ Assets/Scenes/IntFirstHouseScene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a988d62a5ce13374abd9a88378508196
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
