+++ Assets/Scenes/IntFirstShopScene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5a353dfc8c321614392be448047530ea
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
