+++ Assets/Scenes/IntFirstStudioScene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 18975fcc83d4e2f47b52f635b2f4f17a
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
