--- Assets/Scenes/InteriorMapFirstScene.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: a988d62a5ce13374abd9a88378508196
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
