--- Assets/Scenes/InteriorMapShopPoor.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 5a353dfc8c321614392be448047530ea
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
