--- Assets/Scenes/InteriorMapsStudioPoor.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 18975fcc83d4e2f47b52f635b2f4f17a
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
