--- Assets/Scenes/MapFirstScene.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 5062a349a3bc40c4eaa6d4abe2b8f8c7
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
