--- Assets/Scenes/MusicStorePoor.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 5d2bd01b467d3224e83c2eed89fa9fb3
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
