--- Assets/Scripts/GameManager/SceneController.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 31ef820cecd5618428fadad6d505aa98
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
