--- Assets/Scripts/GameManager/SceneController/MusicStorePoorSceneController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: dedf6909e83f3434a808eabdba44a53d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
