+++ Assets/Scripts/SceneController/Exterior.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ba048fbcbbd865747a639fda3a389951
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
