+++ Assets/Scripts/SceneController/Exterior/First.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1f9d97f0ecc81aa4e9947c2abb30e68b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
