+++ Assets/Scripts/SceneController/Exterior/First/ExtFirstSceneController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d390e26bc16c3a546ba5efc197c88411
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
