+++ Assets/Scripts/SceneController/Interior.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ef5f41c6469fc3d4d8fa290a6a0e4dc8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
