+++ Assets/Scripts/SceneController/Interior/First.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c5e87c1d77f5aae4cafbfc0cc60383a6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
