+++ Assets/Scripts/SceneController/Interior/First/IntFirstHouseSceneController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dedf6909e83f3434a808eabdba44a53d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
