+++ Assets/Scripts/SceneController/Interior/First/IntFirstMusicStoreSceneController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2dfc4dd529a1bcf499515d66db754c6b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
