+++ Assets/Scripts/SceneController/Interior/First/IntFirstShopSceneController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e9fe04fc521a9a041a16d31e841de488
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
