+++ Assets/Scripts/Utils/LoadExterior.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 764e0121ae2afc546a0112b84d65dcc0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
