+++ Assets/Scripts/Utils/LoadInterior.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e6e56a251f3a7b0408216aee28c9122c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
