+++ ProjectSettings/EditorBuildSettings.asset
@@ -9,18 +9,18 @@ EditorBuildSettings:
     path: 
     guid: 00000000000000000000000000000000
   - enabled: 1
-    path: Assets/Scenes/InteriorMapFirstScene.unity
+    path: Assets/Scenes/IntFirstHouseScene.unity
     guid: a988d62a5ce13374abd9a88378508196
   - enabled: 1
-    path: Assets/Scenes/InteriorMapShopPoor.unity
+    path: Assets/Scenes/IntFirstShopScene.unity
     guid: 5a353dfc8c321614392be448047530ea
   - enabled: 1
-    path: Assets/Scenes/InteriorMapsStudioPoor.unity
+    path: Assets/Scenes/IntFirstStudioScene.unity
     guid: 18975fcc83d4e2f47b52f635b2f4f17a
   - enabled: 1
-    path: Assets/Scenes/MusicStorePoor.unity
+    path: Assets/Scenes/IntFirstMusicStoreScene.unity
     guid: 5d2bd01b467d3224e83c2eed89fa9fb3
   - enabled: 1
-    path: Assets/Scenes/MapFirstScene.unity
+    path: Assets/Scenes/ExtFirstScene.unity
     guid: 5062a349a3bc40c4eaa6d4abe2b8f8c7
   m_configObjects: {}
